ci(pre-commit): fix helm-docs hook
The helm-docs pre-commit hook is currently pointing out to this repository, which is not correct.

Use https://github.com/norwoodj/helm-docs (which has `.pre-commit-hooks.yaml`).

Furthermore, pin the version (`pre-commit` does not play well with `HEAD`).

Last, fix a broken link to the terrateam docs.

Tested: `pre-commit run -a helm-docs`.